Comparison and Hash can be made much cheaper without string formatting.
